Create README.md
Calculates water usage for irrigation systems. Allows you to create irrigation zones, add plants to those zones or drag plants to from zone to zone, enter runtimes, emitter information, and other data for the calculations. Gives granular calculations per plant and zone down to the the second and up to the year, as well as a grand total for your entire irrigation system. The application is also responsive and will adapt to mobile devices including phones and tablets. Built in Vanilla JavaScript, HTML, and CSS. No DB. Save feature saves it to your browser, serialized in local storage.